ChafaTermInfo: Fix issue in 3-argument control sequences for values >255
Values would be cropped to eight bits.

This only affected CHAFA_TERM_SEQ_BEGIN_SIXELS, but since p1-p3 are
usually set to zero, it went unnoticed.

+++ chafa/chafa-term-info.c
@@ -302,7 +302,7 @@ emit_seq_2_args_uint8 (const ChafaTermInfo *term_info, gchar *out, ChafaTermSeq 
 }
 
 static gchar *
-emit_seq_3_args_uint (const ChafaTermInfo *term_info, gchar *out, ChafaTermSeq seq, guint8 arg0, guint8 arg1, guint8 arg2)
+emit_seq_3_args_uint (const ChafaTermInfo *term_info, gchar *out, ChafaTermSeq seq, guint arg0, guint arg1, guint arg2)
 {
     guint args [3];
